Refactor and simplify the find_token logic
In here, I push the idea of moving into a 1-dimensional world a bit
earlier. Dropping line numbers and columns at first sight. This
simplifies all the math everywhere.

This also patches the issue with the Liquid::Tokenizer tokenizing
incomplete tags in a way that I broke a couple of assumptions.