Introduce single file and whole theme checks & optimize LSP
A check is "single-file" if it only depends on the content of a single file.
A check is "whole-theme" if it depends on the content other files in the theme.

In the context of LSP, we don't need to re-run single-file checks for files that haven't changed.

Also, this commit skips analysis on document open, except on the first one.