Support theme app extensions
Adds support for `extends` to the configuration. This allows us to have
a Theme App Extension specific configuration file that doesn't pull in
checks that don't apply for extensions.

Changes the behaviour of multiple -c args on the command line: checks
have to match all of the specified categories instead of any of them

Rename only_categories to include_categories for consistency with the
CLI help.

If any -c / -x parameters are specified on the command line, they will
completely override what is in the config.