Add `HtmlParsingError` check for reporting errors during HTML parsing
The HTML parser we're using (Gumbo) limits the number of attributes per element to 400, and the maximum depth of the DOM tree to 400 levels. We could bump to limits, or remove the limits, but that seems risky to me to parse without limits.

So instead I chose the approach of reporting the errors.

If any one of those limits is reached, parsing stops, and all HTML offenses on this file are ignored.

Fixes #317