Handle rogue carriage returns
Turns out that everything blows up when we have \r without a followup \n in a file. Nokogiri will handle it and give you a new line, but we didn't and so we ended up with nokogiri telling you something was on line 61 and us not having the newline character and the thing being on line 60.

Fixes #528
Fixes #545