Speed up partial theme checks & add "themeCheck.onlySingleFileChecks"
The difference here is that instead of rerunning all the whole theme
checks on change, as we previously did, we only do that on save and
open when "themeCheck.onlySingleFileChecks" is true. What this means is
that whenever you are typing, only the single file checks will run.

This takes ~10ms on my M1 MBP. Not the best machine to do benchmarks on
but it's 125x faster than what we had before. So I'll take it that 100x+
improvement will transfer to slower machines as well :D

When "themeCheck.onlySingleFileChecks" is on, then we disable all whole
theme checks entirely and only report failures for the open files. This
means that it will always be fast, but it comes at the cost of not
having MissingTranslations, UnusedSnippet, and other checks that span
multiple files.

Fixes #542
Fixes #507