Handle variable lookup as names in UnusedAssign
The `{{ [some_var] }}` syntax is valid Liquid. When this happens, the
tree gives us a Liquid::VariableLookup for the node.value.name (instead
of a string).

To handle those cases, we simply return node.value.name.name.

It's not entirely future proof, but fixing this in Shopify/liquid seemed non-trivial.

Fixes #576