Fixes theme-check-disable when disable comment is followed by another
Kind of weird edge case. If the theme-check-disable (:all) comment was followed
by another comment, then the theme-check-disable (:all) comment was made to only
apply on the range of the followup comment.

The fix is to not do that when theme-check-disable :all is meant to be
applied to the entire document.

Fixes #583