Prevent server from hanging on error
For some reason, sometimes $stderr.puts err.full_message hangs.
And I mean the word used, it _hangs_. It doesn't raise an error.
This is annoying because you end up with a 100% CPU situation instead of
logging the error you're meaning to log.

This PR wraps the error printing into timeouts so that they print and
stop hanging.

(What bullshit...)

Fixes #622